Adapted the presented static content. It is now the same as displayed on the current website. Closes #14.
Updated the contents of pages index/home, info and kontakt. Includes a slight change to template.css where I moved a padding statement from div-class 'section' to 'featured'.
The RWTHGG logo at the bottom of the page now links to the Facebook page of the RWTHGG. Centered the logos.
Set a meaningful website title text and replaced the zend framework default logo.